feat: changed libraries. Tried to add middleware that does not work
