+++ src/application/controller.ts
@@ -1,6 +1,7 @@
 import Elysia from "elysia";
 import authService from "./services/authService";
 import homeService from "./services/homeService";
+import Html from "@kitajs/html"
 
 const controller = () => {
   return new Elysia()
