+++ src/component/BaseHtml.tsx
@@ -1,21 +1,24 @@
-import * as elements from "typed-html";
+import { PropsWithChildren } from "@kitajs/html";
 
-const BaseHtml = ({ children }: elements.Children) => `
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <script src="https://unpkg.com/htmx.org@1.9.5" integrity="sha384-xcuj3WpfgjlKF+FXhSQFQ0ZNr39ln+hwjN3npfM9VBnUskLolQAcN80McRIVOPuO" crossorigin="anonymous"></script>
-  <script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio"></script>
-  <meta name="htmx-config" content='{"globalViewTransitions":true}'>
-  <title>inHouse</title>
-</head>
-<body hx-boost="true">
-  ${children}
-</body>
-</html>
-`;
+const BaseHtml = ({ name, children }: PropsWithChildren<{ name?: string }>) => {
+  return (
+    <html lang="en">
+      <head>
+        <meta charset="UTF-8" />
+        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <script
+          src="https://unpkg.com/htmx.org@1.9.5"
+          integrity="sha384-xcuj3WpfgjlKF+FXhSQFQ0ZNr39ln+hwjN3npfM9VBnUskLolQAcN80McRIVOPuO"
+          crossorigin="anonymous"
+        ></script>
+        <script src="https://cdn.tailwindcss.com?plugins=forms,typography,aspect-ratio"></script>
+        <meta name="htmx-config" content='{"globalViewTransitions":true}' />
+        <title>{name || "inHouse"}</title>
+      </head>
+      <body hx-boost="true">{children}</body>
+    </html>
+  );
+};
 
 export default BaseHtml;
