+++ src/middleware/mainHtml.ts
@@ -1,0 +1,8 @@
+import Elysia from "elysia";
+
+const mainHtml = () => {
+    return new Elysia()
+
+}
+
+export default mainHtml;
