+++ src/pages/SignIn.tsx
@@ -1,9 +1,8 @@
-import * as elements from "typed-html";
 import BaseHtml from "../component/BaseHtml";
 
 const SignIn = () => {
   return (
-    <BaseHtml>
+    <BaseHtml name="SignIn">
       <div class="flex min-h-full flex-col justify-center py-12 sm:px-6 lg:px-8">
         <div class="sm:mx-auto sm:w-full sm:max-w-md">
           <img
